Create field.go and Move fieldConfig to it
